refactor(mcp): extract Studio payload helpers to _studio_payloads (ADR-0008)
Rebasing #1914 onto main (with #1924's studio.py growth merged) pushed studio.py
over the 1000-line ADR-0008 module-size budget. Extract the cohesive wire-shape
projection helpers — _generation_payload / _mind_map_tree / _mind_map_id /
_artifact_rename_payload — into a new _studio_payloads.py sibling (matching the
existing _studio_items / _studio_download splits), bringing studio.py to 934.
Trimmed the studio_generate mind-map docstring to hold SCHEMA_CHAR_BUDGET
(39_382/39_400). Documented the new module in docs/architecture.md.
