+++ tests/unit/mcp/test_tool_eval.py
@@ -31,7 +31,7 @@ pytest.importorskip("fastmcp")
 #: to ~36.0k). Move these DOWN as the surface gets leaner; a rise means
 #: description/param bloat that must be justified, not rubber-stamped.
 SCHEMA_CHAR_BUDGET = (
-    39_400  # total serialized inputSchema + description chars (current 39_379; +21 slack)
+    39_400  # total serialized inputSchema + description chars (current 39_382; +18 slack)
 )
 # #1914 normalized studio_generate's mind-map payload to the bare node tree (behavior,
 # not new params); its docstring note the tree/``null`` shape is +18 chars (39_361 →
